Add tests and CI workflow
